pyproject.toml
[build-system]
# Should be the same as `$POETRY_VERSION` in Dockerfile:
requires = ["poetry==1.2.2"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "django-rest-boilerplate"
version = "1.0.0"
description = ""
authors = ["Patrick Mazulo <pmazulo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9.0,<4.0"
gunicorn = "^20.1.0"
environs = "^9.5.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
uwsgi = "^2.0.21"
django-health-check = "^3.17.0"
python-dateutil = "^2.8.2"
dj-database-url = "^1.0.0"
redis = "4.3.4"
django-storages = "^1.13.1"
boto3 = "^1.25.2"
validators = "^0.20.0"
django-extra-fields = "^3.0.2"
coreapi = "^2.3.3"
django-rest-passwordreset = "^1.3.0"
sentry-sdk = "^1.10.1"
drf-spectacular = "^0.24.2"
importlib-metadata = "^5.0.0"
misspellings-lib = "^2.0.7"
customerio = "^1.5"
django = "^4.1"
werkzeug = "^2.2.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.0"}
psycopg2 = "^2.9.5"


[tool.poetry.group.dev.dependencies]
model-bakery = "1.9.0"
pre-commit = "^2.17.0"
coverage = "^6.5.0"
ipdb = "^0.13.9"
django-debug-toolbar = "^3.7.0"
ipython = "^8.5.0"
flake8 = "^5.0.4"
bandit = "^1.7.4"
pytest = "^7.2.0"
pytest-env = "^0.8.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
black = "^22.10.0"
pytest-xdist = {version = "^3.0.2", extras = ["psutil"]}
pytest-spec = "^3.2.0"
pytest-django = "^4.5.2"
pyinstrument = "^4.3.0"
freezegun = "^1.2.2"
faker = "^15.1.1"
types-requests = "^2.28.11.2"
types-python-dateutil = "^2.8.19.2"
pytest-vcr = "^1.0.2"
pgcli = "^3.5.0"
django-migration-linter = "^4.1.0"
mypy = "^0.961"
django-stubs = {version = "^1.12.0", extras = ["compatible-mypy"]}
djangorestframework-stubs = "^1.7.0"
pudb = "^2022.1.2"


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | _local
  | bin
  | misc
  | _public
  | _data
  | config
  | migrations
)/
'''
